Unexpose UndoBuffer move undo to its own file.
To get undo for a buffer, you now just call NewUndoer(buffer) and you are
returned a buffer with Undo support
